requirements.txt
Pillow>=9.0.0
mcp>=0.1.0
